+++ tree/src/main/java/com/software/shell/tree/TraversalAction.java
@@ -35,4 +35,12 @@ public interface TraversalAction<T extends TreeNode> {
 	 */
 	void perform(T node);
 
+	/**
+	 * Checks whether the traversal is completed and no more required
+	 *
+	 * @return {@code true} if traversal is completed and no more required,
+	 *         {@code false} otherwise
+	 */
+	boolean isCompleted();
+
 }